+++ composeApp/src/commonMain/kotlin/com/pl/myworkoutapp/data/database/WorkoutSessionEntity.kt
@@ -7,8 +7,8 @@ import kotlin.time.Instant
 
 @Entity
 data class WorkoutSessionEntity(
-    @PrimaryKey(autoGenerate = false) val id: Int,
-    val planId: Int,
+    @PrimaryKey(autoGenerate = false) val id: Long,
+    val planId: Long,
     val startTime: Instant?,
     val endTime: Instant?,
     val calories: Double?,
